fix(int64): literales, asignacion, args de fn, listas/maps/JSON y builtins ya no truncan enteros a 32 bits
Fase 1a del plan de deuda. Antes: let a = 3000000000 -> -1294967296; t += 3e9 truncaba;
3000000000 > 2147483647 daba false; json_parse/json()/listas/maps/args de fn perdian el
rango; now_ms() y epoch pasaban por (int). Variable.int_value ya era long long: el corte
estaba en el lexer (atoi/strtol->int), en yylval.ival (int), en ASTNode.value (int) y en
~50 casts (int) al crear literales INT/NUMBER o al guardar en int_value.

- parser.l: strtoll en literales decimales/hex/binarios. parser.y: %union ival long long.
- ast.h/ast.c: ASTNode.value long long; create_ast_leaf/_number/create_int_node.
- (int) -> (long long) en create_ast_leaf_number(INT|NUMBER, ...) (ast.c, te_csv, te_linq,
  te_linq_ops, te_math, te_stdlib, te_string) y en asignaciones a int_value.
- %d -> %lld en todos los printf de value/int_value (ast.c, te_json, debugger, wasm, main).
- 3 prototipos locales desactualizados (typeeasy_api, te_csv, mysql_bridge) corregidos:
  el compilador los detecto como conflicting types.
- test tests/lang/01_types/int64_roundtrip.te (14 casos; falla en 0.0.34, pasa ahora).

Pendiente (Fase 1b): la evaluacion sigue en double -> exacto hasta 2^53; TeValue para i64 pleno.
Regresion Win: lang 98/98 (+3 xfail), api 9/9, ERP main.te syntax-check ok, 39 modulos.

+++ src/ast.h
@@ -68,7 +68,7 @@ typedef struct ASTNode {
     char *type;
     NodeKind kind;        /* Fase 1: cached enum kind */
     char *id;
-    int value;
+    long long value;   /* int64: literales/IDs/timestamps no se truncan a 32 bits */
     char *str_value;
     struct ASTNode *left;
     struct ASTNode *right;
@@ -377,7 +377,7 @@ void free_object_node(ObjectNode *obj);
 ObjectNode *create_objects_bulk(ClassNode *cls, int N);
 void call_method(ObjectNode *obj, char *method);
 ClassNode *find_class(char *name);
-ASTNode *create_int_node(int value);
+ASTNode *create_int_node(long long value);
 ASTNode *create_float_node(int value);
 ASTNode *create_train_node(const char *model_name, const char *dataset_name, ASTNode *options);
 ASTNode *create_train_option_node(const char *option_name, int value);
@@ -422,7 +422,7 @@ ASTNode *create_ast_node_for(char *type, ASTNode *var, ASTNode *init, ASTNode *c
 Variable *find_variable_for(char *id);
 Variable *find_variable(char *id);
 void add_or_update_variable(char *id, ASTNode *value);
-ASTNode *create_ast_leaf(char *type, int value, char *str_value, char *id);
+ASTNode *create_ast_leaf(char *type, long long value, char *str_value, char *id);
 ASTNode *create_ast_node(char *type, ASTNode *left, ASTNode *right);
 ASTNode *create_return_node(ASTNode *expr);
 ASTNode *create_call_node(const char *funcName, ASTNode *args);
@@ -433,7 +433,7 @@ ASTNode *add_statement(ASTNode *list, ASTNode *stmt);
 ASTNode *create_case_node(ASTNode* condition, ASTNode* body);
 ASTNode *append_case_clause(ASTNode* list, ASTNode* case_clause);
 ASTNode *create_match_node(ASTNode* condition, ASTNode* case_list);
-ASTNode *create_ast_leaf_number(char *type, int value, char *str_value, char *id);
+ASTNode *create_ast_leaf_number(char *type, long long value, char *str_value, char *id);
 ASTNode *create_kv_pair_node(char *key, ASTNode *value);
 char* get_node_string(ASTNode *node);
 int is_string_type(ASTNode *node);
